backend/requirements.txt
Flask==2.1.1
tensorflow>=2.10.0
numpy==1.23.0
Pillow==9.2.0